requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langgraph>=0.0.40
langchain-google-genai>=1.0.0
langchain-tavily>=0.2.0
python-dotenv>=1.0.0
langchain-community>=0.0.20
typing-extensions>=4.8.0